+++ directives/spec/spec_runner.js
@@ -1,0 +1,11 @@
+/**
+ * Copyright 2016 aixigo AG
+ * Released under the MIT license.
+ * http://laxarjs.org/license
+ */
+export const title = 'Directives Specification';
+export const tests = [
+   'id_spec',
+   'layout_spec',
+   'widget_area_spec'
+];