+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 
 ## Last Changes
 
+- [#1](https://github.com/LaxarJS/laxar-angular-adapter/issues/1): directives: moved from core to this adapter
+
 
 ## v0.1.0